メール通知等の通知処理追加

+++ front/src/components/organisms/stayroom/StayroomReserveList.vue
@@ -2,6 +2,7 @@
   <div>
     <ConfirmModal v-show="cancelFlg"
       :modal-msg-prop="modalMsg"
+      :is-processing-prop="isCanceling"
       @process-confirm="deleteStayroomReserve"/>
     <div v-for="(reserve, i) in reserves" :key="i">
       <transition name="fade-default" :key="i">
@@ -60,7 +61,8 @@ export default {
         title: "宿泊予約取消",
         message: "宿泊の予約を取り消します。よろしいですか？",
         btn: "取消"
-      }
+      },
+      isCanceling: false
     }
   },
   methods: {
@@ -72,34 +74,43 @@ export default {
     },
     // キャンセル処理の実行
     deleteStayroomReserve(cancelConfirmFlg) {
-      this.cancelFlg = false;
       if (cancelConfirmFlg) {
-        axios.delete(
-          `http://${g.hostName}/api/stayrooms/${this.selecteReserve.stayroom_id}/reserves/${this.selecteReserve.id}`,
-          {
-            headers: {
-              Authorization: this.$store.getters['user/secureToken']
+        if (!this.isCanceling) {
+          this.isCanceling = true;
+          axios.delete(
+            `http://${g.hostName}/api/stayrooms/${this.selecteReserve.stayroom_id}/reserves/${this.selecteReserve.id}`,
+            {
+              headers: {
+                Authorization: this.$store.getters['user/secureToken']
+              }
             }
-          }
-        )
-        .then(() => {
-          this.reserves.splice(this.removeNo, 1)
-          this.$store.dispatch(
-            "flash/create",
-            { message: "取消が完了しました",
-              type:    1
-            }
-          );
-        })
-        .catch((error) => {
-          this.$store.dispatch(
-            "flash/create",
-            { message: error.response.data,
-              type:    2
-            }
-          );
-          this.apiErrors(error.response);
-        });
+          )
+          .then(() => {
+            this.reserves.splice(this.removeNo, 1)
+            this.$store.dispatch(
+              "flash/create",
+              { message: "取消が完了しました",
+                type:    1
+              }
+            );
+          })
+          .catch((error) => {
+            this.$store.dispatch(
+              "flash/create",
+              { message: error.response.data,
+                type:    2
+              }
+            );
+            this.apiErrors(error.response);
+          })
+          .finally(() => {
+            this.isCanceling = false;
+            this.cancelFlg = false;
+          });
+        }
+      }
+      else {
+        this.cancelFlg = false;
       }
     }
   },
